agents/triage-agent/requirements.txt
# Core dependencies
anthropic>=0.40.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# DeepSeek (OpenAI-compatible)
# Uses openai package with custom base_url

# Notion integration
notion-client>=2.0.0

# Database
supabase>=2.0.0

# Jupyter notebook support
ipykernel>=6.0.0
jupyter>=1.0.0

# Data / analysis
numpy>=1.26.0
